# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Message {
  id: number;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  body: string;
  createdAt: string;
}

export interface CreateMessageInput {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  body: string;
}

export interface TopContributor {
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  messageCount: number;
}

export interface ChatStats {
  totalMessages: number;
  messagesToday: number;
  activeMembersToday: number;
  topContributors: TopContributor[];
}

export interface PresenceMember {
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface PresenceSnapshot {
  onlineCount: number;
  members: PresenceMember[];
}

export interface UserProfile {
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface DirectMessage {
  id: number;
  senderId: string;
  recipientId: string;
  senderUsername: string;
  /** @nullable */
  senderAvatarUrl: string | null;
  body: string;
  /** @nullable */
  voiceAudioId: number | null;
  /** @nullable */
  voiceDurationMs: number | null;
  /** @nullable */
  readAt: string | null;
  createdAt: string;
}

export interface SendVoiceMessageInput {
  /**
   * @minLength 1
   * @maxLength 10000000
   */
  audioBase64: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  mimeType: string;
  /**
   * @minimum 1
   * @maximum 600000
   */
  durationMs: number;
}

export interface Status {
  id: number;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  body: string;
  createdAt: string;
  expiresAt: string;
}

export interface CreateStatusInput {
  /**
   * @minLength 1
   * @maxLength 280
   */
  body: string;
}

export interface Conversation {
  partner: UserProfile;
  lastMessage: DirectMessage;
  unreadCount: number;
}

export interface UnreadCount {
  total: number;
}

export interface MarkReadResponse {
  marked: number;
}

export interface MeProfile {
  id: string;
  username: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  avatarUrl: string | null;
  firstSeenAt: string;
  lastSeenAt: string;
  /** @nullable */
  lastLoginAt: string | null;
  /** @nullable */
  lastLogoutAt: string | null;
  loginCount: number;
}

/**
 * Missing or invalid authentication
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

export type ListMessagesParams = {
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
};

export type ListUsersParams = {
  search?: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};