# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: messages
    description: Public lounge messages
  - name: presence
    description: Presence and activity
  - name: users
    description: Discoverable members
  - name: dms
    description: Private direct messages
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List recent messages
      description: Returns the most recent chat messages in chronological order (oldest first)
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Send a chat message
      description: Persists a new chat message and broadcasts it to connected clients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageInput"
      responses:
        "201":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /messages/stats:
    get:
      operationId: getChatStats
      tags: [messages]
      summary: Lounge activity overview
      description: Aggregate stats — total messages, active members today, top contributors in the last 24 hours
      responses:
        "200":
          description: Activity summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /presence:
    get:
      operationId: getPresence
      tags: [presence]
      summary: Currently online members
      description: Lightweight snapshot of who is connected right now
      responses:
        "200":
          description: Online members snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresenceSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List discoverable members
      description: Members the current user can start a private conversation with
      parameters:
        - in: query
          name: search
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a member's profile
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /dms:
    get:
      operationId: listConversations
      tags: [dms]
      summary: List the current user's conversations
      description: Returns one entry per conversation partner with the latest message and unread count
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dms/unread:
    get:
      operationId: getUnreadDmCount
      tags: [dms]
      summary: Total unread direct messages across all conversations
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadCount"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dms/{userId}:
    get:
      operationId: listDirectMessages
      tags: [dms]
      summary: List messages exchanged with a specific user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DirectMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: sendDirectMessage
      tags: [dms]
      summary: Send a direct message to a specific user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageInput"
      responses:
        "201":
          description: Direct message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /dms/{userId}/voice:
    post:
      operationId: sendVoiceMessage
      tags: [dms]
      summary: Send a voice message to a specific user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendVoiceMessageInput"
      responses:
        "201":
          description: Voice message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /statuses:
    get:
      operationId: listStatuses
      tags: [statuses]
      summary: List active statuses (not expired)
      responses:
        "200":
          description: List of active statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Status"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createStatus
      tags: [statuses]
      summary: Post a new status (expires in 24h)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStatusInput"
      responses:
        "201":
          description: Status posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/login:
    post:
      operationId: recordLogin
      tags: [me]
      summary: Record a login event for the current user
      description: Upserts the user's profile in the local database and records a login event
      responses:
        "200":
          description: Login recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/logout:
    post:
      operationId: recordLogout
      tags: [me]
      summary: Record a logout event for the current user
      responses:
        "200":
          description: Logout recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get the current user's saved profile
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dms/{userId}/read:
    post:
      operationId: markConversationRead
      tags: [dms]
      summary: Mark all messages from this user as read
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkReadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Message:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - avatarUrl
        - body
        - createdAt
    CreateMessageInput:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - body
    TopContributor:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        messageCount:
          type: integer
      required:
        - userId
        - username
        - avatarUrl
        - messageCount
    ChatStats:
      type: object
      properties:
        totalMessages:
          type: integer
        messagesToday:
          type: integer
        activeMembersToday:
          type: integer
        topContributors:
          type: array
          items:
            $ref: "#/components/schemas/TopContributor"
      required:
        - totalMessages
        - messagesToday
        - activeMembersToday
        - topContributors
    PresenceMember:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - userId
        - username
        - avatarUrl
    PresenceSnapshot:
      type: object
      properties:
        onlineCount:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/PresenceMember"
      required:
        - onlineCount
        - members
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - userId
        - username
        - avatarUrl
    DirectMessage:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: string
        recipientId:
          type: string
        senderUsername:
          type: string
        senderAvatarUrl:
          type: ["string", "null"]
        body:
          type: string
        voiceAudioId:
          type: ["integer", "null"]
        voiceDurationMs:
          type: ["integer", "null"]
        readAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - senderId
        - recipientId
        - senderUsername
        - senderAvatarUrl
        - body
        - voiceAudioId
        - voiceDurationMs
        - readAt
        - createdAt
    SendVoiceMessageInput:
      type: object
      properties:
        audioBase64:
          type: string
          minLength: 1
          maxLength: 10000000
        mimeType:
          type: string
          minLength: 1
          maxLength: 100
        durationMs:
          type: integer
          minimum: 1
          maximum: 600000
      required:
        - audioBase64
        - mimeType
        - durationMs
    Status:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - avatarUrl
        - body
        - createdAt
        - expiresAt
    CreateStatusInput:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 280
      required:
        - body
    Conversation:
      type: object
      properties:
        partner:
          $ref: "#/components/schemas/UserProfile"
        lastMessage:
          $ref: "#/components/schemas/DirectMessage"
        unreadCount:
          type: integer
      required:
        - partner
        - lastMessage
        - unreadCount
    UnreadCount:
      type: object
      properties:
        total:
          type: integer
      required:
        - total
    MarkReadResponse:
      type: object
      properties:
        marked:
          type: integer
      required:
        - marked
    MeProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        firstSeenAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
        lastLogoutAt:
          type: ["string", "null"]
          format: date-time
        loginCount:
          type: integer
      required:
        - id
        - username
        - email
        - avatarUrl
        - firstSeenAt
        - lastSeenAt
        - lastLoginAt
        - lastLogoutAt
        - loginCount