# lib/api-zod/src/generated/types/createStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStatusInput {
  /**
   * @minLength 1
   * @maxLength 280
   */
  body: string;
}