# lib/api-zod/src/generated/types/directMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DirectMessage {
  id: number;
  senderId: string;
  recipientId: string;
  senderUsername: string;
  /** @nullable */
  senderAvatarUrl: string | null;
  body: string;
  /** @nullable */
  voiceAudioId: number | null;
  /** @nullable */
  voiceDurationMs: number | null;
  /** @nullable */
  readAt: Date | null;
  createdAt: Date;
}