# lib/api-zod/src/generated/types/sendVoiceMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendVoiceMessageInput {
  /**
   * @minLength 1
   * @maxLength 10000000
   */
  audioBase64: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  mimeType: string;
  /**
   * @minimum 1
   * @maximum 600000
   */
  durationMs: number;
}