# lib/api-zod/src/generated/types/status.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Status {
  id: number;
  userId: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
  body: string;
  createdAt: Date;
  expiresAt: Date;
}